ovm-ironic/requirements.txt
python-ironicclient
future==0.15.2
pbr==1.10.0
six==1.10.0
stevedore>=1.16.0
retrying
